pyproject.toml
[project]
name = "tiktok-to-tweets-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anyio>=4.12.1",
    "fastapi>=0.128.0",
    "google-genai>=1.61.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
